requirements.txt
azure-identity==1.15.0
langchain==0.0.329
azure-search-documents==11.4.0b8
python-dotenv==1.0.0
fastapi==0.108.0
uvicorn[standard]==0.25.0
tiktoken==0.5.2
qdrant-client==1.4.0
# Non-coursera version sentence-transformers==2.3.1
sentence-transformers==2.2.2
openai==1.11.1

